+++ pkg/poll/providers/providers.go
@@ -6,6 +6,7 @@
 package providers
 
 import (
+	_ "dns-companion/pkg/poll/providers/caddy"    // Register Caddy provider
 	_ "dns-companion/pkg/poll/providers/docker"   // Register Docker provider
 	_ "dns-companion/pkg/poll/providers/file"     // Register File provider
 	_ "dns-companion/pkg/poll/providers/remote"   // Register Traefik provider