+++ CHANGELOG.md
@@ -7,6 +7,7 @@
      - (poll) Added File provider to read YAML/JSON/Hosts/Zonefile from filesystem with customizable interval to poll for changes or ondemand/fsnotify
      - (poll) Added Remote provider to read YAML/JSON/Hosts/Zonefile from a HTTP/HTTPS source with basic authentication supported
      - (poll) Added Zerotier Poll provider to poll for nodes in a Zerotier Central or ZT-Net (Self hosted) network
+     - (poll) Added Caddy Poll provider to read host.domains from Caddy Admin API
      - (dns) support multiple providers
      - (output) Add functionality to output records to various files (hosts, json, yaml, zone)
      - (output/hosts) auto flatten cnames to accomodate for deficiencies in host file format